+++ anixart/__init__.py
@@ -28,6 +28,5 @@ Usage:
 
 from .__version__ import __license__, __description__
 from .__version__ import __version__, __url__, __build__, __title__, __author__, __author_email__, __copyright__
-from .api import AnixAPIRequests as AnixAPI
-from .api import AnixUserAccount
-from .endpoints import AnixComment, AnixProfileVotedSort
+from .api import AnixUserAccount, AnixAPI
+from .endpoints import AnixComment, AnixProfileVotedSort, AnixList