+++ docs/CHANGELOG.md
@@ -1,6 +1,39 @@
 # CHANGELOG
 
+## TODOs
+
+* **До конца изучить Anixart API.**
+* **Доделать раздел API Anixart в документации.**
+
 ## Anixart API Wrapper
+
+### 22.01.2022
+#### Version: 0.3.5, Build: 540
+
+* Добавлен класс `anix.profile.blocklist` с методами:
+    * get
+    * add
+    * remove
+* Добавлен класс `anix.profile.list` с методами:
+    * get
+* Добавлен класс `anix.coll.favorite` с методами:
+    * get
+    * add
+    * delete
+* В классе `anix.other` были добавлены методы:
+    * history
+    * toggles
+* Метод `anix.profile.blocklist()` перенесён в класс `anix.profile.blocklist` и теперь называется `anix.profile.bl.get(page)`.
+* Оптимизирована инициализация классов через `@property`.
+* Добавлен `logger`, выводит `DEBUG` информацию.
+
+### 21.01.2022
+#### Version: 0.3.4, Build: 519
+
+* Перенёс сессию запросов в `AnixUserAccount`.
+* Был более подробно изучен Anixart API.
+* Были добавлены эндпоинты.
+* Метод `anix.other.voice()` был исправлен на `anix.other.type()`.
 
 ### 20.01.2022
 #### Version: 0.3.3.1, Build: 509